DESIGN OF HYBRID ALGORITHM TO  TSP SOLUTION
Main ant colony optimization algorithms ACO used in several applied fields and on many real life problems with the aim of optimizing the results related to a particular problem, one of the main real life significant problems " Travelling Sales Man Problem (TSP) " and it considers as a crucial problem that have solved using many of algorithms especially smart algorithms (Artificial Intelligence Un-Blind Search Algorithms). Many of the programming environments deals with the applied of this issue solution from the construction of the problem space then define the implementation parts to this problem space components also as a final step determine the solving methodology programming environments that will be Java programming environment. In this research the application of one of the versions MMAS(Min-MAX ANT SYSTEM) and genetic algorithm on TSP problem is structured with an overview of the MACO techniques used in the problem of TSP in the aim finding the best  optimi